Move dialogs into package
