+++ pom.xml
@@ -225,6 +225,7 @@
                 <configuration>
                     <excludes>
                         <exclude>io/blt/gregbot/ui/components/*</exclude>
+                        <exclude>io/blt/gregbot/ui/dialogs/*</exclude>
                         <exclude>io/blt/gregbot/ui/forms/*</exclude>
                     </excludes>
                 </configuration>
