+++ src/main/java/io/blt/gregbot/ui/dialogs/About.form
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<form xmlns="http://www.intellij.com/uidesigner/form/" version="1" bind-to-class="io.blt.gregbot.ui.forms.About">
+<form xmlns="http://www.intellij.com/uidesigner/form/" version="1" bind-to-class="io.blt.gregbot.ui.dialogs.About">
   <grid id="cbd77" binding="contentPane" layout-manager="GridBagLayout">
     <constraints>
       <xy x="48" y="54" width="592" height="297"/>
