+++ src/main/java/io/blt/gregbot/ui/dialogs/SplashScreen.form
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<form xmlns="http://www.intellij.com/uidesigner/form/" version="1" bind-to-class="io.blt.gregbot.ui.forms.SplashScreen">
+<form xmlns="http://www.intellij.com/uidesigner/form/" version="1" bind-to-class="io.blt.gregbot.ui.dialogs.SplashScreen">
   <grid id="cbd77" binding="contentPane" layout-manager="BorderLayout" hgap="0" vgap="0">
     <constraints>
       <xy x="48" y="54" width="517" height="333"/>
